+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ### Feature
 
+- Add nonce and gas to offline [#661](https://github.com/MyEtherWallet/MyEtherWallet/pull/661)
 - Detect when user not in interface and prompt [#684](https://github.com/MyEtherWallet/MyEtherWallet/pull/684)
 
 ### Release v5.0.2